requirements.txt
Flask==2.3.2
requests==2.31.0
beautifulsoup4==4.12.2
scikit-learn==1.3.0
googlesearch-python==1.2.3